+++ src/main/java/com/yuuko/core/commands/audio/commands/LastCommand.java
@@ -5,7 +5,6 @@ import com.yuuko.core.Configuration;
 import com.yuuko.core.commands.Command;
 import com.yuuko.core.commands.audio.AudioModule;
 import com.yuuko.core.commands.audio.handlers.AudioManagerManager;
-import com.yuuko.core.commands.audio.handlers.GuildAudioManager;
 import com.yuuko.core.utilities.MessageHandler;
 import com.yuuko.core.utilities.TextUtility;
 import net.dv8tion.jda.core.EmbedBuilder;
@@ -20,8 +19,7 @@ public class LastCommand extends Command {
 	@Override
 	public void onCommand(MessageReceivedEvent e, String[] command) {
 		try {
-			GuildAudioManager manager = AudioManagerManager.getGuildAudioManager(e.getGuild().getId());
-			AudioTrack track = manager.scheduler.getLastTrack();
+			AudioTrack track = AudioManagerManager.getGuildAudioManager(e.getGuild().getId()).player.getPlayingTrack();
 
 			if(track != null) {
 				String[] uri = track.getInfo().uri.split("=");