+++ src/main/java/com/yuuko/core/commands/developer/DeveloperModule.java
@@ -3,10 +3,7 @@ package com.yuuko.core.commands.developer;
 import com.yuuko.core.CommandExecutor;
 import com.yuuko.core.commands.Command;
 import com.yuuko.core.commands.Module;
-import com.yuuko.core.commands.developer.commands.LavalinkNodeCommand;
-import com.yuuko.core.commands.developer.commands.ReloadApiCommand;
-import com.yuuko.core.commands.developer.commands.SetStatusCommand;
-import com.yuuko.core.commands.developer.commands.SyncGuildsCommand;
+import com.yuuko.core.commands.developer.commands.*;
 import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
 
 public class DeveloperModule extends Module {
@@ -16,7 +13,8 @@ public class DeveloperModule extends Module {
                 new SetStatusCommand(),
                 new SyncGuildsCommand(),
                 new ReloadApiCommand(),
-                new LavalinkNodeCommand()
+                new LavalinkNodeCommand(),
+                new ReloadDatabaseCommand()
         });
 
         if(e == null || e.getAuthor().getIdLong() != 215161101460045834L) {